feat: use contra widget in example
